% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedep.R
\name{get_approx_dup_ids}
\alias{get_approx_dup_ids}
\title{Get Approximate Duplicate IDs}
\usage{
get_approx_dup_ids(
  d,
  id_cols,
  uidName = NULL,
  customIdName = NULL,
  max_dist = 0.1,
  asList = FALSE
)
}
\arguments{
\item{d}{data}

\item{id_cols}{id_cols}

\item{uidName}{uid name}

\item{customIdName}{custom name}

\item{max_dist}{max dist}

\item{asList}{as list}
}
\value{

}
\description{
Get Approximate Duplicate IDs
}
